Refactor: 인터페이스 분리 작업 완료
