+++ src/main/java/me/synology/hajubal/coins/crawler/SiteData.java
@@ -1,0 +1,10 @@
+package me.synology.hajubal.coins.crawler;
+
+public interface SiteData {
+
+    String getSiteName();
+
+    String getDomain();
+
+    String getBoardUrl();
+}
