+++ src/main/java/me/synology/hajubal/coins/service/WebCrawlerService.java
@@ -43,7 +43,7 @@ public class WebCrawlerService {
             try {
                 pointUrlSet.addAll(crawler.crawling());
             } catch (IOException e) {
-                log.error("Crawling fail. site name: {}", crawler.getSiteName());
+                log.error("Crawling fail.");
                 log.error(e.getMessage(), e);
             }
         });
